Clean up some more code

+++ src/org/traccar/protocol/Xt2400ProtocolDecoder.java
@@ -119,7 +119,7 @@ public class Xt2400ProtocolDecoder extends BaseProtocolDecoder {
         Position position = new Position(getProtocolName());
 
         for (byte tag : format) {
-            switch ((int) tag) {
+            switch (tag) {
                 case 0x03:
                     DeviceSession deviceSession = getDeviceSession(
                             channel, remoteAddress, String.valueOf(buf.readUnsignedInt()));
